+++ tx_service/include/cc/local_cc_handler.h
@@ -48,7 +48,7 @@ public:
                       uint32_t key_shard_code,
                       TxNumber tx_number,
                       int64_t tx_term,
-                      uint16_t command_id,
+                      uint32_t command_id,
                       uint64_t ts,
                       bool is_insert,
                       CcHandlerResult<std::vector<AcquireKeyResult>> &hres,
@@ -62,7 +62,7 @@ public:
                          NodeGroupId ng_id,
                          TxNumber txn,
                          int64_t tx_term,
-                         uint16_t command_id,
+                         uint32_t command_id,
                          bool is_insert,
                          CcHandlerResult<AcquireAllResult> &hres,
                          CcProtocol proto,
@@ -75,7 +75,7 @@ public:
                       NodeGroupId ng_id,
                       uint64_t tx_number,
                       int64_t tx_term,
-                      uint16_t command_id,
+                      uint32_t command_id,
                       uint64_t commit_ts,
                       CcHandlerResult<PostProcessResult> &hres,
                       OperationType op_type,
@@ -97,7 +97,7 @@ public:
     /// <param name="is_deleted"></param>
     CcReqStatus PostWrite(uint64_t tx_number,
                           int64_t tx_term,
-                          uint16_t command_id,
+                          uint32_t command_id,
                           uint64_t commit_ts,
                           const CcEntryAddr &ccentry_addr,
                           const TxRecord *record,
@@ -110,7 +110,7 @@ public:
     CcReqStatus PostRead(
         uint64_t tx_number,
         int64_t tx_term,
-        uint16_t command_id,
+        uint32_t command_id,
         uint64_t key_ts,
         uint64_t gap_ts,
         uint64_t commit_ts,
@@ -141,7 +141,7 @@ public:
               ReadType read_type,
               uint64_t tx_number,
               int64_t tx_term,
-              uint16_t command_id,
+              uint32_t command_id,
               const uint64_t ts,
               CcHandlerResult<ReadKeyResult> &hres,
               IsolationLevel iso_level = IsolationLevel::ReadCommitted,
@@ -153,7 +153,7 @@ public:
               bool abort_if_oom = false) override;
 
     void ReadOutside(int64_t tx_term,
-                     uint16_t command_id,
+                     uint32_t command_id,
                      TxRecord &rec,
                      bool is_deleted,
                      uint64_t commit_ts,
@@ -167,7 +167,7 @@ public:
                    ReadType read_type,
                    uint64_t tx_number,
                    int64_t tx_term,
-                   uint16_t command_id,
+                   uint32_t command_id,
                    const uint64_t ts,
                    CcHandlerResult<ReadKeyResult> &hres,
                    IsolationLevel iso_level = IsolationLevel::RepeatableRead,
@@ -190,7 +190,7 @@ public:
                    ReadType read_type,
                    uint64_t tx_number,
                    int64_t tx_term,
-                   uint16_t command_id,
+                   uint32_t command_id,
                    const uint64_t ts,
                    CcHandlerResult<ReadKeyResult> &hres,
                    IsolationLevel iso_level = IsolationLevel::RepeatableRead,
@@ -205,7 +205,7 @@ public:
                   bool inclusive,
                   uint64_t tx_number,
                   int64_t tx_term,
-                  uint16_t command_id,
+                  uint32_t command_id,
                   uint64_t ts,
                   CcHandlerResult<ScanOpenResult> &hd_res,
                   ScanDirection direction = ScanDirection::Forward,
@@ -226,7 +226,7 @@ public:
                        bool inclusive,
                        uint64_t tx_number,
                        int64_t tx_term,
-                       uint16_t command_id,
+                       uint32_t command_id,
                        uint64_t ts,
                        CcHandlerResult<ScanOpenResult> &hd_res,
                        ScanDirection direction = ScanDirection::Forward,
@@ -239,7 +239,7 @@ public:
                        NodeGroupId node_group_id,
                        uint64_t tx_number,
                        int64_t tx_term,
-                       uint16_t command_id,
+                       uint32_t command_id,
                        uint64_t start_ts,
                        const TxKey &start_key,
                        bool start_inclusive,
@@ -264,7 +264,7 @@ public:
                        uint64_t read_ts,
                        uint64_t tx_number,
                        int64_t tx_term,
-                       uint16_t command_id,
+                       uint32_t command_id,
                        CcHandlerResult<RangeScanSliceResult> &hd_res,
                        IsolationLevel iso_level = IsolationLevel::ReadCommitted,
                        CcProtocol proto = CcProtocol::OCC) override;
@@ -323,13 +323,13 @@ public:
     void ReloadCache(NodeGroupId ng_id,
                      TxNumber tx_number,
                      int64_t tx_term,
-                     uint16_t command_id,
+                     uint32_t command_id,
                      CcHandlerResult<Void> &hres) override;
 
     void FaultInject(const std::string &fault_name,
                      const std::string &fault_paras,
                      int64_t tx_term,
-                     uint16_t command_id,
+                     uint32_t command_id,
                      const TxId &txid,
                      std::vector<int> &vct_node_id,
                      CcHandlerResult<bool> &hres) override;
@@ -348,7 +348,7 @@ public:
                          NodeGroupId ng_id,
                          TxNumber tx_number,
                          int64_t tx_term,
-                         uint16_t command_id,
+                         uint32_t command_id,
                          CcHandlerResult<Void> &hres) override;
 
     void BroadcastStatistics(const TableName &table_name,
@@ -357,7 +357,7 @@ public:
                              NodeGroupId ng_id,
                              TxNumber tx_number,
                              int64_t tx_term,
-                             uint16_t command_id,
+                             uint32_t command_id,
                              CcHandlerResult<Void> &hres) override;
 
     void ObjectCommand(const TableName &table_name,
@@ -367,7 +367,7 @@ public:
                        TxCommand &obj_cmd,
                        TxNumber txn,
                        int64_t tx_term,
-                       uint16_t command_id,
+                       uint32_t command_id,
                        uint64_t tx_ts,
                        CcHandlerResult<ObjectCommandResult> &hres,
                        IsolationLevel iso_level,
@@ -384,7 +384,7 @@ public:
     CcReqStatus UploadTxCommands(
         uint64_t tx_number,
         int64_t tx_term,
-        uint16_t command_id,
+        uint32_t command_id,
         const CcEntryAddr &cce_addr,
         uint64_t obj_version,
         uint64_t commit_ts,
@@ -398,7 +398,7 @@ public:
                              bool flush,
                              uint64_t tx_number,
                              int64_t tx_term,
-                             uint16_t command_id,
+                             uint32_t command_id,
                              CcHandlerResult<bool> &hres) override;
 
     /*
@@ -413,7 +413,7 @@ public:
 
     void BlockCcReqCheck(uint64_t tx_number,
                          int64_t tx_term,
-                         uint16_t command_id,
+                         uint32_t command_id,
                          const CcEntryAddr &cce_addr,
                          CcHandlerResultBase *hres,
                          ResultTemplateType type,
@@ -422,7 +422,7 @@ public:
     void BlockAcquireAllCcReqCheck(uint32_t ng_id,
                                    uint64_t tx_number,
                                    int64_t tx_term,
-                                   uint16_t command_id,
+                                   uint32_t command_id,
                                    std::vector<CcEntryAddr> &cce_addr,
                                    CcHandlerResultBase *hres) override;
